+++ common/src/main/java/me/pandamods/fallingtrees/api/TreeHandler.java
@@ -79,11 +79,11 @@ public class TreeHandler {
 	
 	public static Optional<Float> getMiningSpeed(Player player, BlockPos blockPos, float baseSpeed) {
 		TreeSpeed treeSpeed = TREE_SPEED_CACHES.compute(player.getUUID(), (uuid, speed) -> {
-			if (speed == null || speed.blockPos != blockPos) {
+			if (speed == null || !speed.blockPos.equals(blockPos)) {
 				BlockState blockState = player.level().getBlockState(blockPos);
 				TreeType tree = TreeRegistry.getTree(blockState);
 				if (tree == null) return null;
-				TreeData data = tree.gatherTreeData(blockPos, player.level());
+				TreeData data = tree.gatherTreeData(blockPos, player.level(), player);
 				if (data == null) return null;
 				return new TreeSpeed(data.miningSpeedModifier().getMiningSpeed(baseSpeed), blockPos.immutable());
 			}