+++ common/src/main/java/me/pandamods/fallingtrees/trees/GenericTree.java
@@ -17,8 +17,10 @@ import me.pandamods.fallingtrees.api.TreeType;
 import me.pandamods.fallingtrees.config.FallingTreesConfig;
 import me.pandamods.fallingtrees.config.common.tree.GenericTreeConfig;
 import me.pandamods.fallingtrees.exceptions.TreeTooBigException;
+import me.pandamods.pandalib.platform.Services;
 import net.minecraft.core.BlockPos;
 import net.minecraft.core.Direction;
+import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.level.BlockGetter;
 import net.minecraft.world.level.Level;
 import net.minecraft.world.level.block.LeavesBlock;
@@ -41,7 +43,9 @@ public class GenericTree implements TreeType {
 	}
 
 	@Override
-	public TreeData gatherTreeData(BlockPos blockPos, Level level) {
+	public TreeData gatherTreeData(BlockPos blockPos, Level level, Player player) {
+		if (getConfig().requireTool && !getConfig().allowedToolFilter.isValid(player.getMainHandItem())) return null;
+		
 		blockPos = blockPos.immutable();
 		TreeData.Builder builder = TreeData.builder();