+++ common/src/main/java/me/pandamods/fallingtrees/trees/VerticalTree.java
@@ -17,6 +17,7 @@ import me.pandamods.fallingtrees.api.TreeType;
 import me.pandamods.fallingtrees.config.FallingTreesConfig;
 import me.pandamods.fallingtrees.config.common.tree.VerticalTreeConfig;
 import net.minecraft.core.BlockPos;
+import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.level.Level;
 import net.minecraft.world.level.block.state.BlockState;
 
@@ -30,7 +31,9 @@ public class VerticalTree implements TreeType {
 	}
 
 	@Override
-	public TreeData gatherTreeData(BlockPos blockPos, Level level) {
+	public TreeData gatherTreeData(BlockPos blockPos, Level level, Player player) {
+		if (getConfig().requireTool && !getConfig().allowedToolFilter.isValid(player.getMainHandItem())) return null;
+
 		blockPos = blockPos.immutable();
 		TreeData.Builder builder = TreeData.builder();